requirements.txt
django>=4.0,<5.0
djangorestframework>=3.14
psycopg2>=2.9
django-environ>=0.10  